Added namespace to Middleware tests
